fixed minor issues throughout
