+++ client/src/components/AddTeam.jsx
@@ -57,8 +57,8 @@ const AddTeam = () => {
     };
 
     return (
-        <main className="mainContainer">
-            <div className="">
+        <main>
+            <div>
                 <div className="titleContainer">
                     <h4 className="card-header p-2">Add a new team</h4>
                 </div>
